LayoutTests/http/tests/security/frameNavigation/navigation-ALLOWED-popup.html

<html>
<head>
<script src="resources/lib.js"></script>
<script>
navUtils.set_frame_count(5);

navUtils.onready = function () {
  navUtils.log('Navigating frames.');
  navUtils.navigateByLocation(popup0);
  navUtils.navigateByOpen('window1');
  navUtils.navigateByForm('window2');
  navUtils.navigateByHyperlink('window3');
  navUtils.navigateByPlugin('window4');
};

navUtils.ondone = function () {
  navUtils.log('Navigation complete.');
  navUtils.logWindowCount();
  navUtils.log('popup0: ' + navUtils.canAccess(popup0));
  navUtils.log('popup1: ' + navUtils.canAccess(popup1));
  navUtils.log('popup2: ' + navUtils.canAccess(popup2));
  navUtils.log('popup3: ' + navUtils.canAccess(popup3));
  navUtils.log('popup4: ' + navUtils.canAccess(popup4));

  navUtils.log('Done.');
  if (window.layoutTestController)
    layoutTestController.notifyDone();
};
</script>
</head>
<body>
<div id='console'></div>
<script>
var popup0 = window.open("http://localhost:8000/security/frameNavigation/resources/ready.html", "window0");
var popup1 = window.open("http://localhost:8000/security/frameNavigation/resources/ready.html", "window1");
var popup2 = window.open("http://localhost:8000/security/frameNavigation/resources/ready.html", "window2");
var popup3 = window.open("http://localhost:8000/security/frameNavigation/resources/ready.html", "window3");
var popup4 = window.open("http://localhost:8000/security/frameNavigation/resources/ready.html", "window4");
</script>
</body>
</html>
