LayoutTests/http/tests/security/frameNavigation/navigation-ALLOWED-popup-expected.txt

Navigating frames.
Navigation complete.
Window count: 6
popup0: true
popup1: true
popup2: true
popup3: true
popup4: true
Done.
